should be INT(11) not tinyint(4), sms_count limited to 127 ! :)

+++ db/playsms-upgrade-098-to-099.sql
@@ -1,3 +1,3 @@
 ALTER TABLE `playsms_tblUser` ADD `send_as_unicode` TINYINT(4) NOT NULL DEFAULT '0' AFTER `plus_sign_add`;
 
-ALTER TABLE `playsms_tblSMSOutgoing_queue` ADD `sms_count` TINYINT(4) NOT NULL DEFAULT '0' AFTER `flag`;
+ALTER TABLE `playsms_tblSMSOutgoing_queue` ADD `sms_count` INT(11) NOT NULL DEFAULT '0' AFTER `flag`;
